[integration tests] Move Security tests in dedicated package
